prolog json-ast: add typed ast and conversion
